Add golden image tests
