+++ widget/accordioncontainer_test.go
@@ -1,54 +1,56 @@
-package widget
+package widget_test
 
 import (
 	"testing"
 
 	"fyne.io/fyne"
+	"fyne.io/fyne/internal/painter/software"
 	"fyne.io/fyne/test"
 	"fyne.io/fyne/theme"
+	"fyne.io/fyne/widget"
 
 	"github.com/stretchr/testify/assert"
 )
 
 func TestAccordionContainer(t *testing.T) {
-	ai := NewAccordionItem("foo", NewLabel("foobar"))
+	ai := widget.NewAccordionItem("foo", widget.NewLabel("foobar"))
 	t.Run("Initializer", func(t *testing.T) {
-		ac := &AccordionContainer{Items: []*AccordionItem{ai}}
+		ac := &widget.AccordionContainer{Items: []*widget.AccordionItem{ai}}
 		assert.Equal(t, 1, len(ac.Items))
 	})
 	t.Run("Constructor", func(t *testing.T) {
-		ac := NewAccordionContainer(ai)
+		ac := widget.NewAccordionContainer(ai)
 		assert.Equal(t, 1, len(ac.Items))
 	})
 }
 
 func TestAccordionContainer_Empty(t *testing.T) {
-	ac := NewAccordionContainer()
+	ac := widget.NewAccordionContainer()
 	assert.Equal(t, 0, len(ac.Items))
 }
 func TestAccordionContainer_Append(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo", NewLabel("foobar")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo", widget.NewLabel("foobar")))
 	assert.Equal(t, 1, len(ac.Items))
 }
 func TestAccordionContainer_Remove(t *testing.T) {
-	ai := NewAccordionItem("foo", NewLabel("foobar"))
-	ac := NewAccordionContainer(ai)
+	ai := widget.NewAccordionItem("foo", widget.NewLabel("foobar"))
+	ac := widget.NewAccordionContainer(ai)
 	ac.Remove(ai)
 	assert.Equal(t, 0, len(ac.Items))
 }
 func TestAccordionContainer_RemoveIndex(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo", NewLabel("foobar")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo", widget.NewLabel("foobar")))
 	ac.RemoveIndex(0)
 	assert.Equal(t, 0, len(ac.Items))
 }
 
 func TestAccordionContainer_Open(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo0", NewLabel("foobar0")))
-	ac.Append(NewAccordionItem("foo1", NewLabel("foobar1")))
-	ac.Append(NewAccordionItem("foo2", NewLabel("foobar2")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo0", widget.NewLabel("foobar0")))
+	ac.Append(widget.NewAccordionItem("foo1", widget.NewLabel("foobar1")))
+	ac.Append(widget.NewAccordionItem("foo2", widget.NewLabel("foobar2")))
 
 	ac.Open(0)
 	assert.True(t, ac.Items[0].Open)
@@ -70,10 +72,10 @@ func TestAccordionContainer_Open(t *testing.T) {
 }
 
 func TestAccordionContainer_OpenAll(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo0", NewLabel("foobar0")))
-	ac.Append(NewAccordionItem("foo1", NewLabel("foobar1")))
-	ac.Append(NewAccordionItem("foo2", NewLabel("foobar2")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo0", widget.NewLabel("foobar0")))
+	ac.Append(widget.NewAccordionItem("foo1", widget.NewLabel("foobar1")))
+	ac.Append(widget.NewAccordionItem("foo2", widget.NewLabel("foobar2")))
 
 	ac.OpenAll()
 	// Cannot open all items if !accordion.MultiOpen
@@ -90,18 +92,18 @@ func TestAccordionContainer_OpenAll(t *testing.T) {
 }
 
 func TestAccordionContainer_Close(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo", NewLabel("foobar")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo", widget.NewLabel("foobar")))
 	ac.Close(0)
 	assert.False(t, ac.Items[0].Open)
 	assert.False(t, ac.Items[0].Detail.Visible())
 }
 
 func TestAccordionContainer_CloseAll(t *testing.T) {
-	ac := NewAccordionContainer()
-	ac.Append(NewAccordionItem("foo0", NewLabel("foobar0")))
-	ac.Append(NewAccordionItem("foo1", NewLabel("foobar1")))
-	ac.Append(NewAccordionItem("foo2", NewLabel("foobar2")))
+	ac := widget.NewAccordionContainer()
+	ac.Append(widget.NewAccordionItem("foo0", widget.NewLabel("foobar0")))
+	ac.Append(widget.NewAccordionItem("foo1", widget.NewLabel("foobar1")))
+	ac.Append(widget.NewAccordionItem("foo2", widget.NewLabel("foobar2")))
 
 	ac.CloseAll()
 	assert.False(t, ac.Items[0].Open)
@@ -109,283 +111,121 @@ func TestAccordionContainer_CloseAll(t *testing.T) {
 	assert.False(t, ac.Items[2].Open)
 }
 
-func TestAccordionContainer_toggleForIndex(t *testing.T) {
-	ai := NewAccordionItem("foo", NewLabel("foobar"))
-	ac := NewAccordionContainer(ai)
-	assert.False(t, ai.Open)
-	toggle := ac.toggleForIndex(0)
-	toggle()
-	assert.True(t, ai.Open)
-	toggle()
-	assert.False(t, ai.Open)
-}
+func TestAccordionContainer_Layout(t *testing.T) {
+	app := test.NewApp()
+	app.Settings().SetTheme(theme.LightTheme())
 
-func TestAccordionContainerRenderer_MinSize(t *testing.T) {
-	t.Run("Empty", func(t *testing.T) {
-		ac := NewAccordionContainer()
-		acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-		min := acr.MinSize()
-		assert.Equal(t, 0, min.Width)
-		assert.Equal(t, 0, min.Height)
-	})
-	t.Run("Single", func(t *testing.T) {
-		ai := NewAccordionItem("foo", NewLabel("foobar"))
-		t.Run("Open", func(t *testing.T) {
-			ac := NewAccordionContainer()
-			ac.Append(ai)
-			ac.Open(0)
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih := acr.headers[0].MinSize()
-			aid := ai.Detail.MinSize()
-			assert.Equal(t, fyne.Max(aih.Width, aid.Width), min.Width)
-			assert.Equal(t, aih.Height+aid.Height+theme.Padding(), min.Height)
-		})
-		t.Run("Closed", func(t *testing.T) {
-			ac := NewAccordionContainer()
-			ac.Append(ai)
-			ac.Close(0)
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih := acr.headers[0].MinSize()
-			assert.Equal(t, aih.Width, min.Width)
-			assert.Equal(t, aih.Height, min.Height)
-		})
-	})
-	t.Run("Multiple", func(t *testing.T) {
-		ai0 := NewAccordionItem("foo0", NewLabel("foobar0"))
-		ai1 := NewAccordionItem("foo1", NewLabel("foobar1"))
-		ai2 := NewAccordionItem("foo2", NewLabel("foobar2"))
-		t.Run("One_Open", func(t *testing.T) {
-			ac := NewAccordionContainer()
-			ac.Append(ai0)
-			ac.Append(ai1)
-			ac.Append(ai2)
-			ac.Open(0)
-			ac.Close(1)
-			ac.Close(2)
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih0 := acr.headers[0].MinSize()
-			aih1 := acr.headers[1].MinSize()
-			aih2 := acr.headers[2].MinSize()
-			aid0 := ai0.Detail.MinSize()
-			width := fyne.Max(aih0.Width, aid0.Width)
-			width = fyne.Max(width, aih1.Width)
-			width = fyne.Max(width, aih2.Width)
-			assert.Equal(t, width, min.Width)
-			height := aih0.Height
-			height += theme.Padding()
-			height += aid0.Height
-			height += theme.Padding()
-			height += aih1.Height
-			height += theme.Padding()
-			height += aih2.Height
-			assert.Equal(t, height, min.Height)
-		})
-		t.Run("All_Open", func(t *testing.T) {
-			ac := &AccordionContainer{
-				MultiOpen: true,
+	for name, tt := range map[string]struct {
+		multiOpen bool
+		items     []*widget.AccordionItem
+		opened    []int
+	}{
+		"single_open_one_item": {
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+			},
+		},
+		"single_open_one_item_opened": {
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+			},
+			opened: []int{0},
+		},
+		"single_open_multiple_items": {
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+				&widget.AccordionItem{
+					Title:  "B",
+					Detail: widget.NewLabel("2"),
+				},
+			},
+		},
+		"single_open_multiple_items_opened": {
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+				&widget.AccordionItem{
+					Title:  "B",
+					Detail: widget.NewLabel("2"),
+				},
+			},
+			opened: []int{0, 1},
+		},
+		"multiple_open_one_item": {
+			multiOpen: true,
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+			},
+		},
+		"multiple_open_one_item_opened": {
+			multiOpen: true,
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+			},
+			opened: []int{0},
+		},
+		"multiple_open_multiple_items": {
+			multiOpen: true,
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+				&widget.AccordionItem{
+					Title:  "B",
+					Detail: widget.NewLabel("2"),
+				},
+			},
+		},
+		"multiple_open_multiple_items_opened": {
+			multiOpen: true,
+			items: []*widget.AccordionItem{
+				&widget.AccordionItem{
+					Title:  "A",
+					Detail: widget.NewLabel("1"),
+				},
+				&widget.AccordionItem{
+					Title:  "B",
+					Detail: widget.NewLabel("2"),
+				},
+			},
+			opened: []int{0, 1},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			accordion := &widget.AccordionContainer{
+				MultiOpen: tt.multiOpen,
 			}
-			ac.Append(ai0)
-			ac.Append(ai1)
-			ac.Append(ai2)
-			ac.OpenAll()
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih0 := acr.headers[0].MinSize()
-			aih1 := acr.headers[1].MinSize()
-			aih2 := acr.headers[2].MinSize()
-			aid0 := ai0.Detail.MinSize()
-			aid1 := ai1.Detail.MinSize()
-			aid2 := ai2.Detail.MinSize()
-			width := fyne.Max(aih0.Width, aid0.Width)
-			width = fyne.Max(width, fyne.Max(aih1.Width, aid1.Width))
-			width = fyne.Max(width, fyne.Max(aih2.Width, aid2.Width))
-			assert.Equal(t, width, min.Width)
-			height := aih0.Height
-			height += theme.Padding()
-			height += aid0.Height
-			height += theme.Padding()
-			height += aih1.Height
-			height += theme.Padding()
-			height += aid1.Height
-			height += theme.Padding()
-			height += aih2.Height
-			height += theme.Padding()
-			height += aid2.Height
-			assert.Equal(t, height, min.Height)
-		})
-		t.Run("One_Closed", func(t *testing.T) {
-			ac := &AccordionContainer{
-				MultiOpen: true,
+			for _, ai := range tt.items {
+				accordion.Append(ai)
 			}
-			ac.Append(ai0)
-			ac.Append(ai1)
-			ac.Append(ai2)
-			ac.Open(0)
-			ac.Open(1)
-			ac.Close(2)
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih0 := acr.headers[0].MinSize()
-			aih1 := acr.headers[1].MinSize()
-			aih2 := acr.headers[2].MinSize()
-			aid0 := ai0.Detail.MinSize()
-			aid1 := ai1.Detail.MinSize()
-			width := fyne.Max(aih0.Width, aid0.Width)
-			width = fyne.Max(width, fyne.Max(aih1.Width, aid1.Width))
-			width = fyne.Max(width, aih2.Width)
-			assert.Equal(t, width, min.Width)
-			height := aih0.Height
-			height += theme.Padding()
-			height += aid0.Height
-			height += theme.Padding()
-			height += aih1.Height
-			height += theme.Padding()
-			height += aid1.Height
-			height += theme.Padding()
-			height += aih2.Height
-			assert.Equal(t, height, min.Height)
+			for _, o := range tt.opened {
+				accordion.Open(o)
+			}
+
+			window := test.NewWindowWithPainter(accordion, software.NewPainter())
+			window.Resize(accordion.MinSize().Max(fyne.NewSize(150, 200)))
+
+			test.AssertImageMatches(t, "accordion_layout_"+name+".png", window.Canvas().Capture())
+
+			window.Close()
 		})
-		t.Run("All_Closed", func(t *testing.T) {
-			ac := NewAccordionContainer()
-			ac.Append(ai0)
-			ac.Append(ai1)
-			ac.Append(ai2)
-			ac.CloseAll()
-			acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-			min := acr.MinSize()
-			aih0 := acr.headers[0].MinSize()
-			aih1 := acr.headers[1].MinSize()
-			aih2 := acr.headers[2].MinSize()
-			width := aih0.Width
-			width = fyne.Max(width, aih1.Width)
-			width = fyne.Max(width, aih2.Width)
-			assert.Equal(t, width, min.Width)
-			height := aih0.Height
-			height += theme.Padding()
-			height += aih1.Height
-			height += theme.Padding()
-			height += aih2.Height
-			assert.Equal(t, height, min.Height)
-		})
-	})
-}
-
-func TestAccordionContainerRenderer_Layout(t *testing.T) {
-	ai0 := NewAccordionItem("foo0", NewLabel("foobar0"))
-	ai1 := NewAccordionItem("foo1", NewLabel("foobar1"))
-	ai2 := NewAccordionItem("foo2", NewLabel("foobar2"))
-	aid0 := ai0.Detail
-	aid1 := ai1.Detail
-	aid2 := ai2.Detail
-	ac := NewAccordionContainer()
-	ac.Append(ai0)
-	ac.Append(ai1)
-	ac.Append(ai2)
-
-	acr := test.WidgetRenderer(ac).(*accordionContainerRenderer)
-	aih0 := acr.headers[0]
-	aih1 := acr.headers[1]
-	aih2 := acr.headers[2]
-
-	t.Run("All_Closed", func(t *testing.T) {
-		ac.CloseAll()
-		min := ac.MinSize()
-		acr.Layout(min)
-		assert.Equal(t, 0, aih0.Position().X)
-		assert.Equal(t, 0, aih0.Position().Y)
-		assert.Equal(t, min.Width, aih0.Size().Width)
-		assert.Equal(t, aih0.MinSize().Height, aih0.Size().Height)
-		assert.Equal(t, 0, aih1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+theme.Padding(), aih1.Position().Y)
-		assert.Equal(t, min.Width, aih1.Size().Width)
-		assert.Equal(t, aih1.MinSize().Height, aih1.Size().Height)
-		assert.Equal(t, 0, aih2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+2*theme.Padding(), aih2.Position().Y)
-		assert.Equal(t, min.Width, aih2.Size().Width)
-		assert.Equal(t, aih2.MinSize().Height, aih2.Size().Height)
-	})
-	t.Run("One_Closed", func(t *testing.T) {
-		ac.MultiOpen = true
-		ac.Close(0)
-		ac.Open(1)
-		ac.Open(2)
-		min := ac.MinSize()
-		acr.Layout(min)
-		assert.Equal(t, 0, aih0.Position().X)
-		assert.Equal(t, 0, aih0.Position().Y)
-		assert.Equal(t, min.Width, aih0.Size().Width)
-		assert.Equal(t, aih0.MinSize().Height, aih0.Size().Height)
-		assert.Equal(t, 0, aih1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+theme.Padding(), aih1.Position().Y)
-		assert.Equal(t, min.Width, aih1.Size().Width)
-		assert.Equal(t, aih1.MinSize().Height, aih1.Size().Height)
-		assert.Equal(t, 0, aih2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+aid1.MinSize().Height+3*theme.Padding(), aih2.Position().Y)
-		assert.Equal(t, min.Width, aih2.Size().Width)
-		assert.Equal(t, aih2.MinSize().Height, aih2.Size().Height)
-		assert.Equal(t, 0, aid1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+2*theme.Padding(), aid1.Position().Y)
-		assert.Equal(t, min.Width, aid1.Size().Width)
-		assert.Equal(t, aid1.MinSize().Height, aid1.Size().Height)
-		assert.Equal(t, 0, aid2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+aid1.MinSize().Height+aih2.MinSize().Height+4*theme.Padding(), aid2.Position().Y)
-		assert.Equal(t, min.Width, aid2.Size().Width)
-		assert.Equal(t, aid2.MinSize().Height, aid2.Size().Height)
-	})
-	t.Run("One_Opened", func(t *testing.T) {
-		ac.Close(0)
-		ac.Close(1)
-		ac.Open(2)
-		min := ac.MinSize()
-		acr.Layout(min)
-		assert.Equal(t, 0, aih0.Position().X)
-		assert.Equal(t, 0, aih0.Position().Y)
-		assert.Equal(t, min.Width, aih0.Size().Width)
-		assert.Equal(t, aih0.MinSize().Height, aih0.Size().Height)
-		assert.Equal(t, 0, aih1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+1*theme.Padding(), aih1.Position().Y)
-		assert.Equal(t, min.Width, aih1.Size().Width)
-		assert.Equal(t, aih1.MinSize().Height, aih1.Size().Height)
-		assert.Equal(t, 0, aih2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+2*theme.Padding(), aih2.Position().Y)
-		assert.Equal(t, min.Width, aih2.Size().Width)
-		assert.Equal(t, aih2.MinSize().Height, aih2.Size().Height)
-		assert.Equal(t, 0, aid2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aih1.MinSize().Height+aih2.MinSize().Height+3*theme.Padding(), aid2.Position().Y)
-		assert.Equal(t, min.Width, aid2.Size().Width)
-		assert.Equal(t, aid2.MinSize().Height, aid2.Size().Height)
-	})
-	t.Run("All_Opened", func(t *testing.T) {
-		ac.MultiOpen = true
-		ac.OpenAll()
-		min := ac.MinSize()
-		acr.Layout(min)
-		assert.Equal(t, 0, aih0.Position().X)
-		assert.Equal(t, 0, aih0.Position().Y)
-		assert.Equal(t, min.Width, aih0.Size().Width)
-		assert.Equal(t, aih0.MinSize().Height, aih0.Size().Height)
-		assert.Equal(t, 0, aih1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aid0.MinSize().Height+2*theme.Padding(), aih1.Position().Y)
-		assert.Equal(t, min.Width, aih1.Size().Width)
-		assert.Equal(t, aih1.MinSize().Height, aih1.Size().Height)
-		assert.Equal(t, 0, aih2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aid0.MinSize().Height+aih1.MinSize().Height+aid1.MinSize().Height+4*theme.Padding(), aih2.Position().Y)
-		assert.Equal(t, min.Width, aih2.Size().Width)
-		assert.Equal(t, aih2.MinSize().Height, aih2.Size().Height)
-		assert.Equal(t, 0, aid0.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+theme.Padding(), aid0.Position().Y)
-		assert.Equal(t, min.Width, aid0.Size().Width)
-		assert.Equal(t, aid0.MinSize().Height, aid0.Size().Height)
-		assert.Equal(t, 0, aid1.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aid0.MinSize().Height+aih1.MinSize().Height+3*theme.Padding(), aid1.Position().Y)
-		assert.Equal(t, min.Width, aid1.Size().Width)
-		assert.Equal(t, aid1.MinSize().Height, aid1.Size().Height)
-		assert.Equal(t, 0, aid2.Position().X)
-		assert.Equal(t, aih0.MinSize().Height+aid0.MinSize().Height+aih1.MinSize().Height+aid1.MinSize().Height+aih2.MinSize().Height+5*theme.Padding(), aid2.Position().Y)
-		assert.Equal(t, min.Width, aid2.Size().Width)
-		assert.Equal(t, aid2.MinSize().Height, aid2.Size().Height)
-	})
+	}
 }
